pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["Alexey Boriskin <alex@boriskin.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
django-rest-framework = "^0.1.0"
djangorestframework-jsonapi = "^7.1.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
mysqlclient = "^2.2.7"
uritemplate = "^4.1.1"
ipdb = "^0.13.13"
pyyaml = "^6.0.2"
django-filter = "^24.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
